requirements.txt
# Authentication & Security
bcrypt
dash-auth>=2.0.0

# Web Framework & Server
dash>=2.16.0,<3.0.0
dash_bootstrap_components>=1.5.0,<2.0.0
dash_iconify
dash_mantine_components>=0.12.0,<1.0.0
flask>=3.0.0,<4.0.0
flask_login
gunicorn

# Database
pymongo>=4.3

# Data Processing & Analysis
exchange_calendars>=4.0.0
pytz

# API & HTTP
python-dotenv
requests

# Streaming
websocket-client>=1.6.0

# Testing
PyYAML>=6.0
